Desktop/pitchperfect/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
fpdf==1.7.2
mistralai==0.0.3
streamlit
PyPDF2
requests
os